+++ src/notification/service.rs
@@ -2,7 +2,7 @@ use chrono::{DateTime, Utc};
 use serde::{Deserialize, Serialize};
 
 use crate::{
-    notification::{NotificationStatus, NotificationSyncSourceKind},
+    notification::{NotificationSourceKind, NotificationStatus, NotificationSyncSourceKind},
     task::TaskId,
     third_party::integrations::google_calendar::GoogleCalendarEventAttendeeResponseStatus,
 };
@@ -13,7 +13,7 @@ pub struct SyncNotificationsParameters {
     pub asynchronous: Option<bool>,
 }
 
-#[derive(Serialize, Deserialize, Debug, Default, PartialEq, Eq)]
+#[derive(Serialize, Deserialize, Debug, Default, PartialEq, Eq, Clone)]
 pub struct NotificationPatch {
     pub status: Option<NotificationStatus>,
     pub snoozed_until: Option<DateTime<Utc>>,
@@ -23,4 +23,11 @@ pub struct NotificationPatch {
 #[derive(Serialize, Deserialize, Debug, PartialEq)]
 pub struct InvitationPatch {
     pub response_status: GoogleCalendarEventAttendeeResponseStatus,
+}
+
+#[derive(Debug, Serialize, Deserialize)]
+pub struct PatchNotificationsRequest {
+    pub status: Vec<NotificationStatus>,
+    pub sources: Vec<NotificationSourceKind>,
+    pub patch: NotificationPatch,
 }